refactor: position sticky를 위해 계층 변경
